+++ src/app/page.tsx
@@ -1,6 +1,8 @@
 import HeroBanner from "@/components/home/HeroBanner";
 import ServiceFeatures from "@/components/home/ServiceFeatures";
 import CategorySection from "@/components/home/CategorySection";
+import FeaturedProducts from "@/components/home/FeaturedProducts";
+import FlashSaleSection from "@/components/home/FlashSaleSection";
 
 export default function Home() {
   return (
@@ -8,6 +10,8 @@ export default function Home() {
       <HeroBanner />
       <ServiceFeatures />
       <CategorySection />
+      <FeaturedProducts />
+      <FlashSaleSection />
     </>
   );
 }