+++ src/types/product.ts
@@ -1,0 +1,12 @@
+export interface Product {
+  id: string;
+  namebn: string;
+  brand: string;
+  image: string;
+  discount: string;
+  rating: string;
+  reviewCount: string;
+  currentPrice: string;
+  oldPrice: string;
+  href: string;
+}